# perf: Further refinements to pathology CLI calls