--- a/pyluna-pathology/luna/pathology/cli/generate_mask.py
+++ b/pyluna-pathology/luna/pathology/cli/generate_mask.py
@@ -52,7 +52,7 @@ def generate_mask(input_slide_image, input_slide_roi, output_dir, annotation_nam
 
     Args:
         input_slide_image (str): path to slide image (.svs)
-        input_slide_roi (str): path to a halo or other polygonal annotation file
+        input_slide_roi (str): path to a halo or other polygonal annotation file (.xml, .geojson)
         output_dir (str): output/working directory
         annotation_name (str): name of annotation layer to use