Brings a whole-slide ETL down to a few minutes:

1. Generate tiles, 20x, 128px tiles (150k) - 23 sec
2. Detect tissue using otsu's method - 1 min 50 sec
3. Run a Resnet18 inference classifier - 21 sec
4. Generate 6 PNG markup images - 3 min 4 sec

Example notebook showing functionality: https://mskcc.box.com/s/xwzzoqoompfymkcf34cqvb9wexxvdjs3