@aauker 2022-01-21
Thanks for the careful review Doori, I’ll implement the suggested changes. There’s definitely a more efficient way to generate/detect tiles. As for model loading, yes one can use local code, but there’s a small bit of updating needed after #268